WIP: Update Dockerfile for CI build
The Dockerfile would not build because of outdated CMake.

I first tried to install CMake via apt, but this failed because it required libssl1.0.0, and
the container OS only had libssl1.1.0.

Installing a new CMake from Kitware's binary installer worked.

Other changes:

 - Use the anaconda base image continuumio/miniconda3
 - Base the install layer off the smaller "debian:bullseye-slim".

If there are any strange errors with this container, we may need to revert to the full "debian:bullseye".