BUG: label_geometry_measures with uint32 labels
To save an unnecessary clone, we were lazily cloning the data but if the labels are actually uint32, the variable wasn't properly initialized.

For label_stats, add checks that labels are unsigned. Also re-index data in the returned DF by label - it was previously sorted by label, but the pandas index was not updated